Made start on primary gameplay loop
Created a sprite sheet and a player class, along with a main file. Obviously very bare bones and janky (sprite has a funky "stretching") error, and I think the number of frames we have for it is very excessive, as on thirty fps it is too fast. Requires a lot of tweaking, but sets out a basic skeleton for creating our primary gameplay loop.